+++ frontend/mietmiez/src/app/components/UserState/AlreadyLoggedIn.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function AlreadyLoggedIn() {
+  return (
+    <main>
+      <div>
+        <p>Please log out first</p>
+      </div>
+    </main>
+  );
+}
